Fix issue on return request error or HTTP code 204

+++ example_send_subscriptions_inbound_message.py
@@ -18,6 +18,7 @@ parser.add_argument("-s", "--server", help="Address of the server (default=https
 parser.add_argument("username", help="Login")
 parser.add_argument("password", help="Password")
 parser.add_argument("address", help="Destination address")
+parser.add_argument("-p", "--port", help="local port for delivery notification")
 parser.add_argument("-d", "--data_format", help="Type of data used in request, can be url or json (default=url)")
 parser.add_argument("-a", "--accept", help="Type of data used for response, can be url or json (default=url)")
 args = parser.parse_args()
@@ -26,6 +27,10 @@ data_format = "url"
 if args.data_format:
     if (args.data_format == "json"):
         data_format = "json"
+
+port = 7090
+if args.port:
+    port = int(args.port)
 
 header = None
 if 'accept' in locals():
@@ -39,7 +44,7 @@ sms_client = oneapi.SmsClient(args.username, args.password, args.server)
 # example:prepare-message-without-notify-url
 sms = models.SMSRequest()
 sms.address = args.address
-sms.notify_url = 'Any URL'
+sms.notify_url = 'http://{}:{}'.format('localhost', port)
 sms.callback_data = 'Any string'
 sms.filter_criteria = "py_test_"+mod_utils.get_random_alphanumeric_string()
 # ----------------------------------------------------------------------------------------------------
@@ -57,10 +62,21 @@ if not result.is_success():
 print 'Is success = ', result.is_success()
 print 'Resource URL = ', result.resource_url
 
+server = dummyserver.DummyWebWerver(port)
+server.start_wait_and_shutdown(15)
+
+requests = server.get_requests()
+if not requests:
+    print 'No requests received'
+    sys.exit(1)
+
+for method, path, http_body in requests:
+    inbound_notif = oneapi.SmsClient.unserialize_inbound_message(http_body)
+    print inbound_notif
+
 #Few seconds later we can delete the subscription
 time.sleep(10)
 
 sms_client = oneapi.SmsClient(args.username, args.password, args.server)
-
 sms_client.delete_messages_sent_subscription(resource_url)
 # ----------------------------------------------------------------------------------------------------
